pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kustobench"
version = "0.1.0"
description = "A benchmark tool for Azure Data Explorer (ADX) / Eventhouse Kusto clusters"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "azure-kusto-data>=4.6.0",
    "azure-identity>=1.15.0",
    "click>=8.1.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.4.0",
]

[project.scripts]
kustobench = "kustobench.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["kustobench*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
